All the admin CRUD operations work successfully

+++ CalorieTrackerApp/src/main/java/com/revature/controllers/ProfileController.java
@@ -2,15 +2,17 @@ package com.revature.controllers;
 
 import com.revature.dto.ProfileDTO;
 import com.revature.dto.UserDTO;
+import com.revature.models.Profile;
 import com.revature.models.User;
 import com.revature.services.ProfileService;
 import com.revature.services.UserService;
-import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @CrossOrigin
@@ -44,9 +46,9 @@ public class ProfileController {
 
     }
 
-    // THIS WILL ONLY BE ACCESSED BY THE ADMIN
+
     @GetMapping("/{pId}")
-    public ResponseEntity<ProfileDTO> getProfile(@PathVariable Integer pId){
+    public ResponseEntity<ProfileDTO> getProfile(@PathVariable int pId){
 
         // Call the profileService.getProfile() method to return the individual profile.
         ProfileDTO profileDTO = profileService.getProfile(pId);
@@ -54,12 +56,6 @@ public class ProfileController {
         return ResponseEntity.ok(profileDTO);
     }
 
-    @GetMapping
-    public ResponseEntity<ProfileDTO> getProfileByUserId(@PathVariable Integer uId){
-        ProfileDTO profileDTO = profileService.getProfileByUserId(uId);
-
-        return ResponseEntity.ok(profileDTO);
-    }
 
     @PutMapping("/{pId}")
     public ResponseEntity<ProfileDTO> updateProfile(@PathVariable int uId, @PathVariable int pId, @RequestBody ProfileDTO profileDTO){
@@ -77,7 +73,43 @@ public class ProfileController {
 
     @DeleteMapping("/{pId}")
     public ResponseEntity<Void> deleteProfile(@PathVariable int uId, @PathVariable int pId){
+
+        UserDTO userDTO = userService.getUserById(uId);
+
+        if (!userDTO.getRole().equals("USER")) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         profileService.deleteProfile(uId, pId);
         return ResponseEntity.noContent().build();
+    }
+
+
+    // THIS WILL ONLY BE ACCESSED BY THE ADMIN
+    @GetMapping
+    public ResponseEntity<ProfileDTO> getProfileByUserId(@PathVariable Integer uId){
+
+        UserDTO userDTO = userService.getUserById(uId);
+
+        if (!userDTO.getRole().equals("ADMIN")) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
+        ProfileDTO profileDTO = profileService.getProfileByUserId(uId);
+
+        return ResponseEntity.ok(profileDTO);
+    }
+
+    @GetMapping("/all")
+    public ResponseEntity<List<ProfileDTO>> getAllProfiles(@PathVariable Integer uId){
+        UserDTO userDTO = userService.getUserById(uId);
+
+        if (!userDTO.getRole().equals("ADMIN")) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
+        List<ProfileDTO> allProfiles = profileService.getAllProfiles(uId);
+
+        return ResponseEntity.ok(allProfiles);
     }
 }
